Fixes shell buffer overflow for large input specs
The `exec()` function buffers the output of the command executed.
When using a large input spec (over 20k lines), the process is
terminated due to overflowing the buffer.
Increasing the buffer also fixes the issue, but for a generic solution,
`spawn()` pipes the output directly without buffering, and behaves just
the same with the option `shell: true`.